fix: make pathfinder composer 2.x compliant
extract mixed class into dedicated file and address filename typo in order to make the project PSR-4 compliant and work with Composer 2
