+++ app/Model/Pathfinder/AbstractPathfinderModel.php
